extracts backend tree picker functions
