+++ src/components/shared/orgunit_picker/OrgUnitTreePicker.js
@@ -1,68 +1,17 @@
 import React, { useState } from "react";
-import { getInstance } from "d2/lib/d2";
-import { TreeViewWithSearch }  from "bluesquare-components";
-import PluginRegistry from "../../core/PluginRegistry";
+import { TreeViewWithSearch } from "bluesquare-components";
+import { getRootData, getChildrenData, request } from "./orgUnitTreeBackend";
 
 // act as a cache to speed up first levels click
-const loadedOrgUnitsById = {};
-let contractsByOrgUnitId = {};
 
-let currentPeriod = undefined
+let currentPeriod = undefined;
 
 const setPeriod = (argPeriod) => {
-  currentPeriod = argPeriod
-}
-
-const defaultOrgUnitFields =
-  "id,name,ancestors[id,name],children[id,name,ancestors[id,name],children[id,name,ancestors[id,name],children]]";
-
-const withHasChildren = (organisationUnits) => {
-  for (let ou of organisationUnits) {
-    ou.has_children = ou.children && ou.children.length > 0;
-    ou.activeContracts = contractsByOrgUnitId[ou.id];
-    
-    loadedOrgUnitsById[ou.id] = ou;
-  }
-  organisationUnits.sort((a, b) => a.name.localeCompare(b.name));
-  return organisationUnits;
-};
-
-export const getRootData = async (id, type = "source") => {
-  if (Object.keys(contractsByOrgUnitId).length == 0) {
-    const contractService = PluginRegistry.extension("contracts.service");
-    const allContracts = await contractService.findAll();
-    contractsByOrgUnitId = _.groupBy(allContracts, (c) => c.orgUnit.id);
-  }
-
-  const d2 = await getInstance();
-  const api = await d2.Api.getApi();
-  const resp = await api.get("organisationUnits", {
-    filter: "level:eq:1",
-    fields: defaultOrgUnitFields,
-    paging: false,
-  });
-
-  return withHasChildren(resp.organisationUnits);
+  currentPeriod = argPeriod;
 };
 
 export const label = (data) => {
   return data.name;
-};
-
-export const getChildrenData = async (id) => {
-  const loadedOrgUnit = loadedOrgUnitsById[id];
-  if (loadedOrgUnit && loadedOrgUnit.children && loadedOrgUnit.children[0].children) {
-    return withHasChildren(loadedOrgUnit.children);
-  }
-  const d2 = await getInstance();
-  const api = await d2.Api.getApi();
-  const resp = await api.get("organisationUnits", {
-    filter: "parent.id:eq:" + id,
-    fields: defaultOrgUnitFields,
-    paging: false,
-  });
-
-  return withHasChildren(resp.organisationUnits);
 };
 
 const search = (input1, input2, type) => {
@@ -70,16 +19,6 @@ const search = (input1, input2, type) => {
   return [];
 };
 
-const request = async (value, count, source, version) => {
-  const d2 = await getInstance();
-  const api = await d2.Api.getApi();
-  const resp = await api.get("organisationUnits", {
-    filter: "name:ilike:" + value,
-    fields: defaultOrgUnitFields,
-  });
-
-  return withHasChildren(resp.organisationUnits);
-};
 const makeDropDownText = (orgUnit) => {
   return (
     <div>
@@ -92,10 +31,11 @@ const makeDropDownText = (orgUnit) => {
           .map((o) => o.name)
           .join(" > ")}
       </pre>
-      <hr></hr>
+      <hr />
     </div>
   );
 };
+
 const onOrgUnitSelect = (orgUnit) => {
   alert("Selected " + orgUnit.name);
   return [];
@@ -110,8 +50,7 @@ const formatInitialSelectedIds = (selection) => [];
 const formatInitialSelectedParents = (selection) => new Map();
 
 const OrgUnitTreePicker = ({ initialSelection, onChange, period }) => {
-
-  setPeriod(period)
+  setPeriod(period);
 
   const [selectedOrgUnits, setSelectedOrgUnits] = useState(initialSelection);
